+++ include/ccglib/transpose/transpose.h
@@ -12,13 +12,13 @@ public:
   void Run(cu::DeviceMemory &d_input, cu::DeviceMemory &d_output);
 
 private:
-  unsigned int B;
-  unsigned int M;
-  unsigned int N;
-  unsigned int M_chunk;
-  unsigned int N_chunk;
+  size_t B;
+  size_t M;
+  size_t N;
+  size_t M_chunk;
+  size_t N_chunk;
 
-  unsigned int nr_bits;
+  size_t nr_bits;
 
   cu::Device &device;
   cu::Stream &stream;
